# src/main/webapp/openapi/swagger.json
openapi: 3.0.3
info:
  title: The Queue
  version: 1.0.0
  description: "API Documentation used for the backend of our exam project"
  contacts: [
{
name: Emil Elkjær Nielsen,
email: cph-en93@cphbusiness.dk
},
{
    name: Sigurd Arik Gaarde Nielsen,
    email: cph-at89@cphbusiness.dk
}
]
  license: {
    name: MIT,
    url: "https://github.com/codergram/TheQueue_Backend/blob/main/LICENSE"
  }
servers:
  - url: "https://api.codergram.me"
    description: Production
paths:
...
  /question:
    put:
      summary: Add answer to question
      description: Add answer to question
      operationId: addAnswerToQuestion
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                answer:
                  type: object
                  properties:
                    answer:
                      type: string
                      example: Et eller andet svar
                    teacher:
                      type: object
                      properties:
                        email:
                          type: string
                          example: underviser@cphbusiness.dk
                    teacherLink:
                      type: string
                      example: "https://badlink.nu"
                id:
                  type: number
                  example: 6
            example:
              answer:
                answer: Et eller andet svar
                teacher:
                  email: underviser@cphbusiness.dk
                teacherLink: "https://badlink.nu"
              id: 6
      responses:
        "200":
          description: Add answer to question
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: "337"
            Date:
              schema:
                type: string
                example: "Wens, 08 June 2022 18:07:36 GMT"
            Keep-Alive:
              schema:
                type: string
                example: timeout=20
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: object
                    properties:
                      answer:
                        type: string
                        example: Et eller andet svar
                      id:
                        type: number
                        example: 6
                      teacher:
                        type: object
                        properties:
                          email:
                            type: string
                            example: underviser@cphbusiness.dk
                          id:
                            type: number
                            example: 1
                          password:
                            type: string
                            example: kodeord123
                          teacherName:
                            type: string
                            example: Emil tutor
                          zoom_url:
                            type: string
                            example: "https://mitnyezoomlink.dk"
                      teacherLink:
                        type: string
                        example: "https://badlink.nu"
                  id:
                    type: number
                    example: 6
                  semesterId:
                    type: number
                    example: 0
              examples:
                Add answer to question:
                  value:
                    answer:
                      answer: Et eller andet svar
                      id: 6
                      teacher:
                        email: underviser@cphbusiness.dk
                        id: 1
                        password: kodeord123
                        teacherName: Emil tutor
                        zoom_url: "https://mitnyezoomlink.dk"
                      teacherLink: "https://badlink.nu"
                    id: 6
                    semesterId: 0
    post:
      summary: Add new question
      description: Add new question
      operationId: addNewQuestion
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: Giver ingen mening
                password:
                  type: string
                  example: "12345"
                semesterId:
                  type: number
                  example: 1
                student:
                  type: object
                  properties:
                    email:
                      type: string
                      example: cph-en93@cphbusiness.dk
                studentLink:
                  type: string
                  example: "http://facebook.ru"
                topic:
                  type: string
                  example: Fejl i postman
            example:
              description: Giver ingen mening
              password: "12345"
              semesterId: 1
              student:
                email: cph-en93@cphbusiness.dk
              studentLink: "http://facebook.ru"
              topic: Fejl i postman
      responses:
        "200":
          description: Add new question
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: "300"
            Date:
              schema:
                type: string
                example: "Wens, 08 June 2022 18:00:14 GMT"
            Keep-Alive:
              schema:
                type: string
                example: timeout=20
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                    example: Giver ingen mening
                  id:
                    type: number
                    example: 10
                  password:
                    type: string
                    example: "12345"
                  semesterId:
                    type: number
                    example: 1
                  student:
                    type: object
                    properties:
                      email:
                        type: string
                        example: cph-en93@cphbusiness.dk
                      id:
                        type: number
                        example: 1
                      name:
                        type: string
                        example: Emil
                  studentLink:
                    type: string
                    example: "http://facebook.ru"
                  timestamp:
                    type: string
                    example: "2022-05-11 20:00:14.431462"
                  topic:
                    type: string
                    example: Fejl i postman
              examples:
                Add new question:
                  value:
                    description: Giver ingen mening
                    id: 10
                    password: "12345"
                    semesterId: 1
                    student:
                      email: cph-en93@cphbusiness.dk
                      id: 1
                      name: Emil
                    studentLink: "http://facebook.ru"
                    timestamp: "2022-05-11 20:00:14.431462"
                    topic: Fejl i postman
...